19/08/23 work

+++ src/Device/securityController.ts
@@ -1,4 +1,5 @@
 import {
+  BadRequestException,
   Controller,
   Delete,
   Get,
@@ -17,7 +18,7 @@ export class SecurityController {
   @UseGuards(RefreshTokenGuard)
   @Get('/devices')
   async getDevices(@Req() req, @Res() res) {
-    const userId = req.user.userId;
+    const userId = req.refreshTokenPayload.userId;
     const devices = await this.devicesService.getDevices(userId);
 
     return res.send(devices).status(200);
@@ -25,16 +26,20 @@ export class SecurityController {
   @UseGuards(RefreshTokenGuard)
   @Delete('/devices')
   async deleteDevices(@Req() req, @Res() res) {
-    const userId = req.user.userId;
-    const deviceId = req.user.deviceId;
+    const userId = req.refreshTokenPayload.userId;
+    const deviceId = req.refreshTokenPayload.deviceId;
+    console.log(userId, deviceId);
     await this.devicesService.deleteDevicesUserExceptForHim(userId, deviceId);
 
     return res.sendStatus(204);
   }
 
   @UseGuards(RefreshTokenGuard)
-  @Delete('/:devices')
+  @Delete('/devices/:deviceId')
   async deleteDevice(@Param('id') deviceId: string, @Req() req, @Res() res) {
+    if (!deviceId) {
+      throw new BadRequestException('id device not found/ control,dev');
+    }
     await this.devicesService.deleteDevice(deviceId);
 
     return res.sendStatus(204);
